requirements.txt
flask
Flask-Dataset
PyYAML
nose
mock
